frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Frontend for the Telegram bot"
readme = "README.md"
authors = [
    { name = "EliaTomaselli", email = "elia.tomaselli@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.9",
]
